requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.19.0
gspread==6.0.0
google-auth==2.28.0
openpyxl==3.1.2